requirements.txt
Django==3.1.4
beautifulsoup4==4.9.3
include
requests==2.24.0
urllib3==1.25.8
django-imagekit==4.0.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
pillow
fake-useragent